:rewind: revert call method to http post, add api handle
