+++ nonebot/adapters/mirai/__init__.py
@@ -1,1 +1,3 @@
 from .bot import MiraiBot
+from .event import *
+from .message import MessageChain, MessageSegment
